+++ yazi-config/src/vfs/service.rs
@@ -39,6 +39,10 @@ pub struct ServiceSftp {
 	pub key_file:       PathBuf,
 	pub key_passphrase: Option<String>,
 	#[serde(default)]
+	pub cert_file:      PathBuf,
+	#[serde(default)]
+	pub no_cert_verify: bool,
+	#[serde(default)]
 	pub identity_agent: PathBuf,
 }
 
@@ -47,6 +51,11 @@ impl ServiceSftp {
 		if !self.key_file.as_os_str().is_empty() {
 			self.key_file = normalize_path(mem::take(&mut self.key_file))
 				.ok_or_else(|| io::Error::other("key_file must be either empty or an absolute path"))?;
+		}
+
+		if !self.cert_file.as_os_str().is_empty() {
+			self.cert_file = normalize_path(mem::take(&mut self.cert_file))
+				.ok_or_else(|| io::Error::other("cert_file must be either empty or an absolute path"))?;
 		}
 
 		self.identity_agent = if self.identity_agent.as_os_str().is_empty() {